Revise: polish up the room_symbol and room_properties form/dialogue files
There were some constructs (`QXxxxLayout`s) that were not attached to
a parent `QWidget` or similar class instances. There were also a few more
of such layouts than there needed to be. In combination this is likely to
make the dynamic resizing of the forms/dialogues not as smooth or as
functional as would be wished.

Also there are a couple of buttons that provide a reset functionality and
those are better done by `QToolButton`s rather than `QPushButton`s.

Within the two files I have also renamed some elements in a more logical/
structured way with the lowercase classname without a q prefix being
first and (for layout elements) the parent class second. Where a
`QGridLayout` was being used but it only had a single column or row of
contents I have also converted them to the more manageable
`Q(V|H)BoxLayouts`.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>