Revise: tweak margins and sizing behaviour in room_properties.ui
Reduces to zero some margins that had previously been at their default
(of 9) but which might not have been utilised in prior code that did not
have them assigned to a `QWidget` and which are now reducing the space
for actual widgets in the form.

I also realised that the sizing of some labels in the form/dialogue because
they had multiple lines of "Engineering English" was making them larger
than they needed to be in use and was compromising the most effective
sizing. By shortening the text (which gets replaced when the form/dialogue
is used); changing their policy to "MinimumExpanding" and also changing
the SizeConstraints of the layouts around them to "SetMininmumSize" I think
they will take up less space.

Also, converted the layout of whole dialogue to be `QVBoxLayout`, it was a
`QGridLayout` but doesn't need to be.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>