Revert: remove most of the new dictionary language codes from OpenBSD
I have left some in place, specifically, these that do not require an
additional translation:
* `cy` -> "Welsh"
+ `cy_gb` -> "Welsh (United Kingdom {Wales})"
+ `fr_xx_classique` -> "French"
+ `vi_x_old`-> "Vietnamese (DauCu variant - old-style diacritics)"
+ `vi_x_new` -> "Vietnamese (DauMoi variant - new-style diacritics)"

and these that still seem reasonably possible to me in Europe:
+ `ga` -> "Irish"
+ `la` -> "Latin"
+ `lb` -> "Luxembourgish"
+ `lb_lu` -> "Luxembourgish (Luxembourg)"

Given that OpenBSD has potentially a large number of other dictionaries
that are not covered by this - and that these unhandled ones can also
occur in distributions other than the Debian related one that I use I
intend to plan out and eventually implement a secondary way to provide
the translation of Hunspell `.aff` + `.dic` file names to localisable
strings via an additional lookup in a read-only `.ini` type file that
OS Distribution packagers (FreeBSD, OpenBSD, GNU/Hurd and other GNU
/Linuxes) can create for THEIR specific distributions to support the files
that they provide. Thereby shifting the translation efforts to them for
the locales and languages that THEY cover.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>