Infrastructure: correct an error associated with `QColor` in "Room properties" and "Room symbol" classes (#6421)
In a couple of places a `QColor` in the `dlgRoomSymbol` class - which
was duplicated in the new `dlgRoomProperties` one - was being compared
to a `nullptr` to determine if it has been set to anything. This is
incorrect because the item will have a memory address whether it has
been assigned a meaningful colour value or not. Instead the
`QColor::isValid()` method is the correct way to do this. (An invalid or
uninitiated QColor has red, green, blue and alpha values all being 0 out
of 255 - i.e. it would be transparent!)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>
Co-authored-by: Kebap <kebap_spam@gmx.net>