fix timestamps being wrongly enabled (#6413)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
File for persistent timestamps was being created every session because
of error console, so second session would come up with timestamps on.
Now only main console saves changes.

#### Motivation for adding to Mudlet

#### Other info (issues closed, discussion etc)
With #6355 I overlooked where consoles other than the main console would
enable timestamps and then they too would have it save to file. The
dlgTriggerEditor constructor was enabling timestamps on the error
console. So now I made it only save that when it is the main console.

I noticed that the error console is not currently showing timestamps,
but that'd be a separate issue. They show in PTB of Sep 1
`ptb-2022-09-01-42d5a` but stopped showing in PTB of Sep 2
`ptb-2022-09-02-e82a2`.

would fix #6408