Infrastructure: centralise predefined MUD game data
Moves the "description" data into the same container as all the other
details for the predefined MUDS. That data is then refactored out of the
`./src/mudlet.h` file into a separate smaller file that only needs to be
`#include`d into two class files - so changing any details requires less
to rebuild the application.

This PR also removes the 3rd party dependency "qt-ordered-map" (which is
3-clause BSD licensed but NOT currently mentioned in our "About" dialogue!)
Afterwards the details for each predefined MUD is held in a QList<T>
instead - with a couple of helper methods defined to reproduce the
information retrieval methods used in the prior code.

This PR also adds an icon for the Mudlet self-test profile which eliminates
a warning in development versions at least about a missing icon for that
"game" - as such that particular profile is NOT select-able with a mouse
and scrolling through the list with the arrow keys causes an adjacent
profile to be highlighted in error!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>