add scrollTo and getScroll lua function (#6406)
#### Brief overview of PR changes/additions
Adds functions for scrolling via script.

#### Motivation for adding to Mudlet
The [bfind script](https://forums.mudlet.org/viewtopic.php?t=1515) on the forums searches through the buffer and shows the found and surrounding lines, but can not offer an echoLink to jump to it.  Scrolling is done with mouse, keys, and the built-in search, but not a way to scroll with scripting yet.  Aside from searches one could also set up scrolling through a text with timers, or you could have a custom key to scroll at specific increments.

#### Other info (issues closed, discussion etc)
Closes #5384
A Lua script utilizing this capability would allow someone to create a custom search rather than complicating the built in search as requested in #6386 (although, that still might be worth doing)

Use `scrollTo(42)` to jump to line 42 in the buffer.  Use `scrollTo(-5)` to get 5 lines back from the end of the buffer, negatives are an equivalent shortcut for `scrollTo(getLastLineNumber()-5)`.  And `scrollTo()` or high number out of range to go to the bottom, or `scrollTo(0)` for the top.  Optionally specify a window in front like `scrollTo("chat",50)`.

Then there is `getScroll([windowName])` to find current value.

Then there are scrollUp and scrollDown, so users can do `scrollUp(2)` instead of `scrollTo(getScroll()-2)`.