Revert: undo changes to GeyserReposition.lua
The lua code put in there was trying to use this new event to force a
redraw of Geyser widgets when a profile hidden during multi-playing when
multi-view was not in use. This was an attempt to solve #5960. However it
seems the problem was with faulty reading of the `TMainConsole` size whilst
it was being shown. That appears to have been solved by a different PR:
#6438 that is waiting to be merged into the code-base. As such the tweaking
with the lua file will not be required.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>