Fix: background color of wrapped text indent
We were using default constructed `TChar` instances in both
`(int) TBuffer::wrap(int)` and other places but these would be constructed
with a black background and a white foreground - and that is not correct
if the background has been set to any other color. By changing the
constructor that previously took a `Host*` to one that takes the `TConsole`
that owns the `TBuffer` where the `TChar` is being used it is possible to
detect and use the current background and foreground color (and text
rendering settings) instead.

There are other places where such default constructed `TChar`s were being
used - it is not immediately obvious whether they actually have any effect
but it seemed sensible to "fix" them as well and see whether it makes any
other changes...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>