Improve: allow `isActive(...)` and `exists(...)` to use an ID number (#6401)
#### Brief overview of PR changes/additions
With this PR passing an actual number as the first argument will use it
as an ID number (provided it rounds to a zero or greater integer) and
return a `1` or `0` if the corresponding Mudlet item is "active" or not
/ exists or not (though like the unmodified one that returns a count of
items of the selected type that are active it does not consider whether
all the item's "ancestors" are active as well before including an item
in the count).

This can be used to identify the active status of temporary items (which
are only identified by an ID number) - provided of course that they are
around long enough to be checked.

N. B. `isActive(...)` will return `0` **if the item (when given an ID
number) or items (when given a name) does NOT exist** - this might not
be what is expected but is how the previous "name-only" form worked.
OTOH ID numbers less than zero are never (:crossed_fingers:) going to
occur and those will provoke a `nil` + error message if used.

The same principle has now been applied to `exists(...)` which will now
return `1` or `0`  if the item with an matching positive ID  number exists
- whether it is a temporary or permanent item.  The numeric value is
needed (rather than a boolean one) so that the result is in the same
form as that when used with a name (string) first argument.

#### Motivation for adding to Mudlet
This came about as a result of a comment in the Discord \#help channel:
> Kalrykh — 2022/10/29 at 15:15 UTC:
>
> I've spent the past few(maybe more than a few) hours setting up some
things with tempTriggers and tempTimers with regular triggers. Is there
a way to show all active temp stuff? i can clear everything out by
cycling mudlet, but I'd like to see if I've been running around
generating tempTriggers/Timers without realizing it due to sloppy code
on my part. if i don't know, they'll just keep getting generated by the
regular triggers i'm using even after cycling mudlet.

> SlySven — 2022/10/29 at 1526 UTC:
> Humm, it seems that `isActive(itemName, itemType)` only works with a
name (string) and not an ID number - so there is no way to find out the
state of a temporary item.

#### Other info (issues closed, discussion etc)
This ought to help with the sort of issue that the OP raised (#6397)
though a script using this revised function would, I think, be needed to
do what is wanted there.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>