Add: command line options to prioritise a predefined MUD and to undo it
The new options are `-o`/`--only` `profileName` - and using it will, like
the `--profile profileName` option launch the identified profile. If it is
one of the "Predefined" ones it also stores a setting in the global Mudlet
settings that mean all **other** Predefined profiles do not get shown from
then on-wards whenever the "Connect Profile" dialogue is shown.

To undo this hiding action a `-a`/`--all` command line option (which does
take priority should they both be given) will reset the stored setting.

Whilst the setting *is* present in the stored settings it is also not
possible for the user to "hide" the Predefined Mud in the dialogue - it
will just not bring up the profile deletion confirmation dialog and
silently not do anything.

Also:
* make `-p` also work alongside `--profile`
* remove several multiple instances of the same `qsl(`...`)` wrapper in
`main.cpp`,  `dlgConnectionProfile`
* eliminate some redundant calls to `(QSetting*) mudlet::getQSettings()`
as the pointer it returns is public there is no need for a "getter".
* rename the badly misnamed `(void) dlgConnectionProfiles::setProfileIcon()`
to `createNonPredefinedProfileItems()`...!
* remove a duplicate call to `(QSetting*) mudlet::getQSettings()` in the
`mudlet` class constructor.
* Handled failures in the command line argument handling by reporting the
error (it didn't previously) and showing the "help" text. In the case of
an error it will also pass a negative (-1) exit value back to the Operating
System/command interpreter. Requesting the help information directly will
continue to return a zero exit value.
* ALL of the help text is now passed through the translation system.
Previously only some of the text was so processed.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>