Add: sysProfileFocusChangeEvent (#6267)
Originally an attempt to close #5960 by detecting when the currently
active host changes when multi-playing and issuing an event to replace
the ones lost when I revised another part of the codebase to NOT
issue a NAWS report or a `sysWindowResizeEvent` when the size of the
main window was NOT actually changed but just shown/hidden. Subsequent
other work revealed that that issue had a different cause. Instead this
PR was expanded to help with #6273, #6449 and #6450 and generally clean
up on matters relating to switching between profiles.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>