Fix: limit maximum replay speed up multiplier
By repeatedly clicking the speed up button it is possible to overflow the
`(int) mudlet::mReplaySpeed` variable such that it takes on a value that is
impossible to work with and then it crashes Mudlet. This PR limits the
maximum (and minimum) multiplier so that it is clamped between 1 and 1,024.

This should close #6466.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>