add support for multiple profiles from command line (#6470)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
This makes it so that `--profile="one" --profile="two"` will connect to
both profiles instead of just the last one.
Also added is `-p "one"` style, and an error message if bad args given.
#### Motivation for adding to Mudlet
Normally I have Mudlet open with five profiles open on one copy for
games that I monitor, and then a second copy of test builds open for
connecting to things I'm testing. If I put a checkmark on "Open profile
on Mudlet start" for each the five profiles, then my test builds will
open up with those same profiles. So, I made a shortcut to open the
release version with the command line arguments listing my 5 profiles.
Only one actually opened. I looked up the process, and it doesn't ask
parser for list of arguments and run through them, but rather asks about
each type of argument. I looked up what the
parser[.value(x)](https://doc.qt.io/qt-5/qcommandlineparser.html#value)
did, and scrolled down and found
[.values(x)](https://doc.qt.io/qt-5/qcommandlineparser.html#values-1)
just below it which is a similar thing using list. It was quicker to
program this change than it was to look that up.

#### Other info (issues closed, discussion etc)
The existing process was to go through the profile folder in
alphabetical order and check if each profile has the box checked or was
named in the `--profile="X"`. ~~I haven't changed that, I only changed
the "equal to X" to "contained in X". So `--profile="one"
--profile="two"` is not going to open `one` and then `two` because of
them being listed in that order, it's going to go through the directory
listing in alphabetical order just like the checkmark way has worked. I
didn't make it run through the list of args in order or anything like
that.~~ Now it will first look at the ones from the command line in the
order given and open them if they are in the folder list and skip it if
they are not (to avoid typos making a new profile), and then for the
remaining folders, it will alphabetically look for the ones with the
autologin set as before.

I'm not aware of any issues from other people who expected the command
line argument to work multiple times. The help prior to my edit already
calls it "additional profile to open", which makes me think that it was
intended to be for opening multiple.

Co-authored-by: Stephen Lyons <slysven@virginmedia.com>