Improve: allow slower than normal replays, and pausing, rewind and stopping
Restructure the replay system to hold all the replay chunks in memory
and to pick them out as required during a replay.

Removes the separate timer for replay time display by inserting dummy
chunks into the above replay data which have an integer number of seconds
as a time but no data. The time information is sent to the display in the
replay toolbar as it is read and since there is a real or dummy chunk at
least every simulated replay second it automatically keeps in step with
the replay elapsed time - however fast it is being replayed.

A pause/resume control has been added which stops the timer used to provide
the delay between replaying each chunk.

A rewind control has been added which pauses the replay and then rewinds
things to the start ready to be re-run by clicking the resume button.

A replay abort (eject) has been added which immediately terminates the
replay and resets the application to run another one.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>