Fix: stop "announcer" widget from masking part of the main menubar
This had been noticed by users/developers even when they were not used
a screen-reader and it was unclear why the "Game" menu item was not
reachable and the "Options" item was hard to get to. The solution is to
"hide" the widget which seems to fix the problem but not otherwise interfer
with the operation of the screen-reading.

This issue was definetely present in the current Windows development
code and I presume it is also an issue for other OSes but I haven't
confirmed it but I have duplicated it in the Unix version.

Also,  there is a `#define` in place in the `./src/AnnouncerWindows.cpp`
file that was being included unconditionally - but it is not needed in ALL
Windows build environments...! (My MSYS2/Mingw-w64 has it correctly defined
already) so I've made it a conditional inclusion.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>