Fix: Reused media players to default & clear media button (#6481)
#### Brief overview of PR changes/additions
When media files are changed in the media cache, it is intended that the media content from the cache will be played. It was discovered that for a media file that was swapped out with the same name, the media player was holding onto the previously loaded file in memory. A change was made such that when a media player is called from the pool of objects that it will be reset to defaults before being reused. Additionally, upon discussing the issue with @gesslar, it was requested that a button to clear media from the cache be added to the profile preferences UI - so this is also done in this PR.

To test this PR, one could download/play a media file `lua playSoundFile({name = "lightning.wav", url = "https://webserver"})` then push the new Clear media button found in the Special options tab of the profile preference, then replace the media file on the webserver with another file with the same name, then download/play the sound again to show it has changed successfully.
#### Motivation for adding to Mudlet
@gesslar identified the issue and requested the new button
#### Other info (issues closed, discussion etc)
Closes #6477

Release notes: A button was added to the profile preferences to clear the profile's cache of media and an issue was fixed to reset media players to defaults upon reuse.