Improve: ensure whole item is selected in room properties QComboBoxes
By ensuring that the whole line of text is selected when the user selects
an (editable) item in the Room Symbol or Weight `QComboBox` it means that
the default action if they type a text entry key will be to delete the
existing text (including any sort of "(count: X room(s)" suffix to the
actual value). This UI improvement will make the entering a new value
easier as they will not have to manually delete the existing entry first.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>