Revise: select all text in any room properties widget & focus on wanted one
Now the focus is put upon the appropriate widget depending on what the
action that was clicked on the 2D mapper context menu - and all the text
there will be pre-selected so that the first key stroke can remove it ready
for a new entry.

Since the "Lock..." and "Unlock..." actions now both end up in the same
place they are duplicates of each other and can be combined.

For clarity, since I am editing some texts in this area I have taken the
liberty to introduce some numerous texts {i.e. `Qbject::tr(...)` calls with
a third numeric argument (in these cases the number of rooms selected)} so
that the texts can be presented as appropriate for a locale when the number
of rooms that the actions will work on varies.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>