Fix: OS dependent DEFINEDs for Screen-readers
For OSes other than MacOs, Windows, Linux and OpenBSD the source code would
fail to compile because some OS dependent portions did not accommodate
them.

This proved to be a problem when I want back to boot into FreeBSD!

It is now set so that the /src/AnnouncerUnix.cpp` file is included for all
OSes other than MacOs or Windows.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>