Fix: crash when reorganising scripts in editor tree view
This will close #6313 which was introduced by #4461. Note that although
it now prevents the crashes - from trying to use a `nullptr` returned
from `QAbstractItemModel::model()` when the `QModelIndex::index` that it
is called for is *invalid* the behaviour is still the same as from the
previously used (but now declared obsolete) `QModelInded::child(int, int,
QModelIndex&)` in that when a (script in this case) is dragged to the
extreme left of the `QTreeWidget` it stays there in a horizontal position
that would normally only be used by the root node in the tree which we
take steps to hide completely in our `TTreeWidget`s.

On saving and loading such cases however the item does get shifted over to
the right to a "normal" position - which is how our code used to behave.
This is still something that should be addressed - so that we intercept
such cases and fix them as they arrive, however my understanding of how
Qt's Model/Views system works is just not good enough for me to be able to
solve that issue currently.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>