Revert: undo/remove feature that moved focus to appropriate control
As requested by @Kebap.

The revisions that adjusted some tool-tips and menu items to include %n
elements or at least have the number provided to the `QObject::tr(...)`
calls to reflect the number of rooms involved have been retained because
otherwise the texts would have had mentions of "room(s)" without any means
for the translators to handle pluralisation in their locale.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>