Revise: fix focus being directed to the main toolbar
For some inexplicable reason the code to switch between multiple loaded
profiles seemed to redirect the focus to buttons on the main toolbar that
have drop-down menus - instead of the wanted command-line - though there is
no indication of an explicit focus policy to do this (maybe it is a built-
in one); this seems to be sufficient to counter it.

There is one down-side in that it is not possible to navigate to the main
toolbar now with `<CTRL>+<TAB>` - it may be desirable to include it in the
sequence that that key-press combination processes.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>