Revise: change the focus modification of some main toolbar buttons
Careful studying of the current `development` code with GammaRay revealed
that the ordinary buttons on the Main Toolbar had the `Qt::NoFocus` policy
and that the ones with a drop down menu had `Qt::TabFocus` - this suggests
that I should have changed the policy in the prior commit for the latter
to also be `Qt::NoFocus` as well (instead of `Qt::ClickFocus`).

Note: it would still be desirable to allow the end-user to `<CTRL>+<TAB>`/
`<CTRL>+<BACKTAB>` to the main toolbar - which I think would require
enabling `Qt::TabFocus` on ALL the main toolbar buttons when the first or
last profile is the active one AND also arrange for it to be be disabled/
overridden when the appropriate key-event happens on the end two buttons
on the Toolbar - however the latter will require coding an event-handler
for that toolbar which is too much to do here and should be considered as
a separate issue.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>