Fix: unknown element erasure
This should give more useful feed-back when unexpected elements are
encountered in game save files and Mudlet item imports from a file or
clipboard. It refactors multiple near-copies of some reporting code to a
common method and extends the functionality to also cover Variables in such
files (which can be set to be saved from the "Variables" view in the
Editor). It also reports the content of the unexpected element better as
it does NOT get caught up with reporting the white-space characters that we
leave between elements in our XML files rather than the meaningful, but
unexpected content.

Rename to better reflect the name of the XML element they parse (though
both "Item" and "ItemGroup" are both parsed by the same method "readItem"):
* `(void) XMLimport::readHostPackage(Host*)`
                                           ==> `XMLimport::readHost(Host*)`
* `(int) XMLimport::readTriggerGroup(TTrigger*)`
                                    ==> `XMLimport::readTrigger(TTrigger*)`
* `(int) XMLimport::readTimerGroup(TTimer*)`
                                        ==> `XMLimport::readTimer(TTimer*)`
* `(int) XMLimport::readAliasGroup(TAlias*)`
                                        ==> `XMLimport::readAlias(TAlias*)`
* `(int) XMLimport::readActionGroup(TAction*)`
                                      ==> `XMLimport::readAction(TAction*)`
* `(int) XMLimport::readScriptGroup(TScript*)`
                                      ==> `XMLimport::readScript(TScript*)`
* `(int) XMLimport::readKeyGroup(TKey*)`    ==> `XMLimport::readKey(TKey*)`

* `(void) XMLimport::readVariableGroup(TVar*)`
                                       ==> `XMLimport::readVariable(TVar*)`

Delete, and replace with calls to unifying:
 `(void) XMLimport::readUnknownElement(const QString&)`:

* `(void) XMLimport::readUnknownPackage()`
* `(void) XMLimport::readUnknownHostElement()`
* `(void) XMLimport::readUnknownTriggerElement()`
* `(void) XMLimport::readUnknownTimerElement()`
* `(void) XMLimport::readUnknownAliasElement()`
* `(void) XMLimport::readUnknownActionElement()`
* `(void) XMLimport::readUnknownScriptElement()`
* `(void) XMLimport::readUnknownKeyElement()`

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>